Update README.md
INTRODUCTION TO BIOINFORMATICS
Bioinformatics has emerged as a scientific discipline that encompasses the application of computing science and technology to analyze and manage biological data. All this began when it was demonstrated by Ingram that there is homology between sickle cell haemoglobin and normal haemoglobin. This led to comparison of other proteins with similar biological function. As more and more proteins were sequenced, it became necessary to have databases which enabled a quick comparison using computational softwares. With the advent of rapid nucleic acid sequencing techniques, a large number of sequences started accumulating which again required computing facilities.
NAME OF CONTRIBUTORS:
1.MUSTAPHA MUKHTAR SAID